AdcApi
