+++ src/http.cpp
@@ -14,6 +14,7 @@
 #include <vector>
 
 #include "adc.hpp"
+#include "api/adc_api.hpp"
 #include "api/commands_api.hpp"
 #include "command_manager.hpp"
 #include "config_manager.hpp"
@@ -29,8 +30,6 @@
 static httpd_handle_t configServer = nullptr;
 static bool fallbackHandlersRegistered = false;
 
-static CommandsApi commands_api(commandManager);
-
 namespace {
 // cppcheck-suppress syntaxError
 extern const char common_css_start[] asm("_binary_common_css_start");
@@ -41,8 +40,6 @@ extern const char common_js_start[] asm("_binary_common_js_start");
 extern const char root_html_start[] asm("_binary_root_html_start");
 // cppcheck-suppress syntaxError
 extern const char status_html_start[] asm("_binary_status_html_start");
-// cppcheck-suppress syntaxError
-extern const char adc_html_start[] asm("_binary_adc_html_start");
 // cppcheck-suppress syntaxError
 extern const char config_html_start[] asm("_binary_config_html_start");
 // cppcheck-suppress syntaxError
@@ -58,55 +55,9 @@ extern const char metrics_html_start[] asm("_binary_metrics_html_start");
 // cppcheck-suppress syntaxError
 extern const char logs_html_start[] asm("_binary_logs_html_start");
 
-#if defined(EBUS_INTERNAL)
-#endif
-
 void sendStatic(httpd_req_t* req, const char* contentType, const char* data) {
   // HttpUtils::sendResponse(req, "200 OK", contentType, std::string(data));
   HttpUtils::sendResponse(req, "200 OK", contentType, data);
-}
-
-uint32_t parseAdcArg(httpd_req_t* req, const char* key, uint32_t fallback) {
-  if (req == nullptr || key == nullptr) return fallback;
-
-  char query[256] = {};
-  if (httpd_req_get_url_query_str(req, query, sizeof(query)) != ESP_OK)
-    return fallback;
-
-  char value[32] = {};
-  if (httpd_query_key_value(query, key, value, sizeof(value)) != ESP_OK)
-    return fallback;
-
-  char* end = nullptr;
-  unsigned long parsed = std::strtoul(value, &end, 10);
-  if (end == value || *end != '\0') return fallback;
-  return static_cast<uint32_t>(parsed);
-}
-
-uint32_t parseAdcChannelMask(httpd_req_t* req) {
-  if (req == nullptr) return 0x03;  // default GPIO0, GPIO1
-
-  char query[256] = {};
-  if (httpd_req_get_url_query_str(req, query, sizeof(query)) != ESP_OK)
-    return 0x03;
-
-  char value[128] = {0};
-  if (httpd_query_key_value(query, "channels", value, sizeof(value)) != ESP_OK)
-    return 0x03;
-
-  uint32_t mask = 0;
-  const char* p = value;
-  while (*p != '\0') {
-    char* end = nullptr;
-    long ch = std::strtol(p, &end, 10);
-    if (end == p) break;
-    if (ch >= 0 && ch <= 4) mask |= (1U << ch);
-    if (*end == ',')
-      p = end + 1;
-    else
-      break;
-  }
-  return mask == 0 ? 0x03 : mask;
 }
 
 esp_err_t handleCommonCss(httpd_req_t* req) {
@@ -258,101 +209,6 @@ esp_err_t handleStatusLib(httpd_req_t* req) {
   return ESP_OK;
 }
 #endif
-
-esp_err_t handleAdcPage(httpd_req_t* req) {
-  sendStatic(req, "text/html", adc_html_start);
-  return ESP_OK;
-}
-
-esp_err_t handleAdcRaw(httpd_req_t* req) {
-  if (!adc.isRunning() && !adc.begin()) {
-    HttpUtils::sendErrorResponse(req, "500 Internal Server Error", "adc_raw",
-                                 "ADC not running or failed to start");
-    return ESP_OK;
-  }
-
-  const uint32_t sampleRate = parseAdcArg(req, "sample_rate", 30000);
-  const uint32_t samplesPerChannel = parseAdcArg(
-      req, "samples_per_channel", parseAdcArg(req, "sample_count", 2400));
-  const uint32_t channelMask = parseAdcChannelMask(req);
-  const uint32_t effectivePerChannelRate =
-      adc.effectivePerChannelSampleRate(sampleRate, channelMask);
-  const uint32_t activeChannelCount =
-      static_cast<uint32_t>(__builtin_popcount(channelMask & 0x1F));
-  const uint32_t controllerRate =
-      effectivePerChannelRate *
-      (activeChannelCount == 0 ? 1U : activeChannelCount);
-
-  const uint64_t captureStartMillis =
-      static_cast<uint64_t>(esp_timer_get_time() / 1000ULL);
-
-  char tmp1[32], tmp2[32], tmp3[32], tmp4[32], tmp5[32], tmp6[32];
-  httpd_resp_set_status(req, "200 OK");
-  httpd_resp_set_type(req, "application/octet-stream");
-  httpd_resp_set_hdr(req, "Cache-Control", "no-cache");
-  httpd_resp_set_hdr(req, "X-ADC-Format", "esp32c3-ch12c3-le16");
-  HttpUtils::applyCustomHeaders(req);
-
-  std::snprintf(tmp1, sizeof(tmp1), "%u",
-                static_cast<unsigned>(effectivePerChannelRate));
-  httpd_resp_set_hdr(req, "X-ADC-Sample-Rate", tmp1);
-
-  std::snprintf(tmp2, sizeof(tmp2), "%u",
-                static_cast<unsigned>(samplesPerChannel));
-  httpd_resp_set_hdr(req, "X-ADC-Samples", tmp2);
-
-  std::snprintf(tmp3, sizeof(tmp3), "%u", static_cast<unsigned>(channelMask));
-  httpd_resp_set_hdr(req, "X-ADC-Channel-Mask", tmp3);
-
-  std::snprintf(tmp4, sizeof(tmp4), "%u",
-                static_cast<unsigned>(Adc::result_bytes));
-  httpd_resp_set_hdr(req, "X-ADC-Result-Bytes", tmp4);
-  std::snprintf(tmp5, sizeof(tmp5), "%llu",
-                static_cast<unsigned long long>(captureStartMillis));
-  httpd_resp_set_hdr(req, "X-ADC-Capture-Start-Millis", tmp5);
-  std::snprintf(tmp6, sizeof(tmp6), "%u",
-                static_cast<unsigned>(controllerRate));
-  httpd_resp_set_hdr(req, "X-ADC-Controller-Sample-Rate", tmp6);
-
-  if (!adc.streamRaw(
-          [req](std::string_view chunk) {
-            httpd_resp_send_chunk(req, chunk.data(), chunk.size());
-          },
-          sampleRate, samplesPerChannel, channelMask))
-    return ESP_FAIL;
-  httpd_resp_send_chunk(req, nullptr, 0);
-  return ESP_OK;
-}
-
-esp_err_t handleAdcState(httpd_req_t* req) {
-  httpd_resp_set_type(req, "application/json;charset=utf-8");
-  HttpUtils::applyCustomHeaders(req);
-  {
-    ebus::detail::JsonWriter writer([req](std::string_view chunk) {
-      httpd_resp_send_chunk(req, chunk.data(), chunk.size());
-    });
-    auto scope = writer.objectScope();
-    writer.writeField("running", adc.isRunning());
-  }
-  httpd_resp_send_chunk(req, nullptr, 0);
-  return ESP_OK;
-}
-
-esp_err_t handleAdcEnable(httpd_req_t* req) {
-  if (adc.begin()) {
-    HttpUtils::sendSuccessResponse(req, "adc_enable");
-  } else {
-    HttpUtils::sendErrorResponse(req, "500 Internal Server Error", "adc_enable",
-                                 "ADC enable failed");
-  }
-  return ESP_OK;
-}
-
-esp_err_t handleAdcDisable(httpd_req_t* req) {
-  adc.stop();
-  HttpUtils::sendSuccessResponse(req, "adc_disable");
-  return ESP_OK;
-}
 
 #if defined(EBUS_INTERNAL)
 esp_err_t handleCronPage(httpd_req_t* req) {
@@ -677,6 +533,7 @@ void SetupHttpHandlers() {
   RegisterUri("/", HTTP_GET, handleRoot);
   RegisterUri("/config", HTTP_GET, handleConfigPage);
   RegisterUri("/upgrade", HTTP_GET, handleUpgradePage);
+  RegisterUri("/api/v1/wifi/scan", HTTP_POST, handleWifiScan);
 
   RegisterUri("/status", HTTP_GET, handleStatusPage);
   RegisterUri("/api/v1/status", HTTP_GET, handleStatus);
@@ -685,14 +542,11 @@ void SetupHttpHandlers() {
   RegisterUri("/api/v1/status/lib", HTTP_GET, handleStatusLib);
 #endif
 
-  RegisterUri("/adc", HTTP_GET, handleAdcPage);
-  RegisterUri("/api/v1/adc/raw", HTTP_GET, handleAdcRaw);
-  RegisterUri("/api/v1/adc/enable", HTTP_POST, handleAdcEnable);
-  RegisterUri("/api/v1/adc/disable", HTTP_POST, handleAdcDisable);
-  RegisterUri("/api/v1/adc/state", HTTP_GET, handleAdcState);
-  RegisterUri("/api/v1/wifi/scan", HTTP_POST, handleWifiScan);
+  static AdcApi adc_api(adc);
+  adc_api.registerHandlers(configServer);
 
 #if defined(EBUS_INTERNAL)
+  static CommandsApi commands_api(commandManager);
   commands_api.registerHandlers(configServer);
 
   RegisterUri("/cron", HTTP_GET, handleCronPage);
